fix(wrong links after refactor): All tests passing
