+++ __tests__/routes/users/users.get.integration.test.ts
@@ -10,7 +10,7 @@ import { RegionRepository } from '../../../src/lib/repositories/RegionRepository
 import routes from '../../../src/lib/routes';
 import {
   DefaultRegions,
-} from '../../../src/lib/types-interfaces';
+} from '../../../src/lib/common';
 import { Region } from '../../../src/orm/entity/Region';
 import {
   closeTestingConnections,
