+++ __tests__/routes/users/users.put.integration.test.ts
@@ -5,7 +5,7 @@ import request from 'supertest';
 import { Connection, getCustomRepository } from 'typeorm';
 import { UserRepository } from '../../../src/lib/repositories/UserRepository';
 import routes from '../../../src/lib/routes';
-import { DefaultRegions, UserRole } from '../../../src/lib/types-interfaces';
+import { DefaultRegions, UserRole } from '../../../src/lib/common';
 import {
   closeTestingConnections,
   createTestingConnections,
