+++ __tests__/routes/users/users.spots.delete.integration.test.ts
@@ -5,7 +5,7 @@ import request from 'supertest';
 import { Connection, getRepository } from 'typeorm';
 import { ERRORS, SUCCESS, SUGGESTIONS } from '../../../src/lib/messages';
 import routes from '../../../src/lib/routes';
-import { UserRole } from '../../../src/lib/types-interfaces';
+import { UserRole } from '../../../src/lib/common';
 import { User } from '../../../src/orm/entity/User';
 import {
   closeTestingConnections,
