+++ src/lib/request-handlers/bathingspots/public-get.ts
@@ -3,7 +3,7 @@ import { Bathingspot } from '../../../orm/entity/Bathingspot';
 import { SUCCESS } from '../../messages';
 import { getRegionsList } from '../../utils/custom-repo-helpers';
 import { RegionRepository } from '../../repositories/RegionRepository';
-import { getResponse, HttpCodes } from '../../common/types-interfaces';
+import { getResponse, HttpCodes } from '../../common';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
 import { BathingspotRepository } from '../../repositories/BathingspotRepository';
