+++ src/lib/request-handlers/responders.ts
@@ -11,8 +11,8 @@ import {
   ResponderSuccessCreated,
   SuccessResponder,
   SuggestionResponder,
-} from '../common/types-interfaces';
-import { ResponderWrongIdOrSuccess } from '../common/types-interfaces';
+} from '../common';
+import { ResponderWrongIdOrSuccess } from '../common';
 
 /**
  * build a payload. This is to reduce repetition
