+++ src/lib/request-handlers/users/post.ts
@@ -4,7 +4,7 @@ import { Region } from '../../../orm/entity/Region';
 import { User } from '../../../orm/entity/User';
 import { getRegionsList } from '../../utils/custom-repo-helpers';
 import { UserRepository } from '../../repositories/UserRepository';
-import { HttpCodes, IObject, postResponse, UserRole } from '../../common/types-interfaces';
+import { HttpCodes, IObject, postResponse, UserRole } from '../../common';
 import { getEntityFields } from '../../utils/get-entity-fields';
 import { errorResponse, responder, responderMissingBodyValue, responderSuccessCreated } from '../responders';
