+++ src/lib/utils/predictions-helpers.ts
@@ -1,5 +1,5 @@
 import { getRepository } from 'typeorm';
-import { IObject } from '../common/types-interfaces';
+import { IObject } from '../common';
 import { BathingspotPrediction } from './../../orm/entity/BathingspotPrediction';
 
 export const createPredictionWithValues: (obj: IObject) => Promise<BathingspotPrediction> = async (obj) => {
