+++ src/setup/populate-db.ts
@@ -1,6 +1,6 @@
 import ora = require('ora');
 import { createConnection, getRepository, getConnectionOptions } from 'typeorm';
-import { DefaultRegions, IAddEntitiesToSpotOptions, UserRole } from '../lib/common/types-interfaces';
+import { DefaultRegions, IAddEntitiesToSpotOptions, UserRole } from '../lib/common';
 import { addEntitiesToSpot } from '../lib/utils/bathingspot-helpers';
 import { Region } from '../orm/entity/Region';
 import { User } from '../orm/entity/User';
